service/requirements.txt
docutils==0.12
lockfile==0.10.2
psutil==2.2.1
python-daemon==2.0.5
